+++ changelog.md
@@ -2,6 +2,8 @@
 
 * Fix connection popup showing blank text sometimes
 * Fix unlisted levels being joinable through the player list in some situations
+* Fix friends not showing at the top of the player list anymore
+* Add new consent notice when first using the mod (will be shown to everyone after updating)
 * Hopefully fix some SSL issues
 
 ## v1.8.1